Renamed `ProcessRunnerException` to `SilentProcessRunnerException` and moved some plumbing bits into a subfolder
