+++ source/SilentProcessRunner/Plumbing/AssemblyInfo.cs
